requirements.txt
tornado
stem
websocket-client
PySocks
requests
